+++ summarizer/parser.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 import os
+import re
 import json
 from collections import defaultdict
 
@@ -8,6 +9,7 @@ from nltk.tokenize import punkt
 
 class SentenceTokenizer(punkt.PunktSentenceTokenizer):
     """ Extend the nltk's punkt sentence tokenizer """
+    _reAbbr = re.compile('((?:[\w]\.)+[\w]*\.)')
     def _annotate_tokens(self, tokens):
         """ Given a set of tokens augmented with markers for line-start and
         paragraph-start, returns an iterator through those tokens with full
@@ -31,8 +33,41 @@ class SentenceTokenizer(punkt.PunktSentenceTokenizer):
         when it is not in the training data, even though they are relatively simple
         to tease out, especially when mixing it with ortho heuristics to detect
         the likelyhood of it being a sentence starter as well an abbreviation."""
-        for token in tokens:
-            yield token
+        for aug_tok1, aug_tok2 in punkt._pair_iter(tokens):
+            if self._reAbbr.search(aug_tok1.tok) is None:
+                yield aug_tok1
+                continue
+
+            aug_tok1.abbr = True
+            aug_tok1.sentbreak = False
+            # Is it the last token? We can't do anything then.
+            if not aug_tok2:
+                continue
+
+            next_typ = aug_tok2.type_no_sentperiod
+            tok_is_initial = aug_tok1.is_initial
+            # figure out if it's a sentence starter
+            # [4.2. Token-Based Reclassification of Abbreviations] If
+            # the token is an abbreviation or an ellipsis, then decide
+            # whether we should *also* classify it as a sentbreak.
+            if (aug_tok1.abbr or aug_tok1.ellipsis) and not tok_is_initial:
+                # [4.1.1. Orthographic Heuristic] Check if there's
+                # orthogrpahic evidence about whether the next word
+                # starts a sentence or not.
+                is_sent_starter = self._ortho_heuristic(aug_tok2)
+                if is_sent_starter == True:
+                    aug_tok1.sentbreak = True
+                    yield aug_tok1
+                    continue
+
+            # [4.1.3. Frequent Sentence Starter Heruistic] If the
+            # next word is capitalized, and is a member of the
+            # frequent-sentence-starters list, then label tok as a
+            # sentence break.
+            if aug_tok2.first_upper and next_typ in self._params.sent_starters:
+                aug_tok1.sentbreak = True
+
+            yield aug_tok1
 
 class Parser(object):
     def __init__(self, ideal=20.0, stop_words=None, tokenizer=None):