+++ CHANGES.md
@@ -1,6 +1,14 @@
 CHANGES
 =======
 
+v0.0.3 10-01-2015
+-----------------
+
+* [FIX] Tokenizer would think a custom abbreviation, F.B.I., was a sentence break
+* [FIX] Added sanitizer module to preprocess text before summarizing it
+* [TESTS] Added ability to pull valid tokenized articles from brevity.detroitnow.io
+and test that the new tokenizer is still valid
+
 v0.0.2 08-26-2015
 -----------------